requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
email_validator==2.1.0
razorpay==1.4.1
